requirements.txt
selenium
pytest
webdriver-manager
python-dotenv
allure-pytest
pytest-xdist
